Added AWS service to store cover image URLs

+++ backend/src/controllers/book.controllers.js
@@ -1,6 +1,8 @@
 import jwt from 'jsonwebtoken';
 import validator from 'validator';
+import s3 from '../services/s3.js';
 import { query } from '../config/db.js'
+import { PutObjectCommand } from '@aws-sdk/client-s3';
 async function createBook(req,res){
     try{
         const {title,description} = req.body;
@@ -29,6 +31,53 @@ async function createBook(req,res){
 
 async function addCover(req,res){
     try{
+        if(!req.file){
+            return res.status(400).json({
+                message: "Cover Image is required"
+            })
+        }
+        const file = req.file;
+
+
+        const fileName = `${Date.now()}-${file.originalname}`;
+
+        const params = {
+            Bucket: process.env.AWS_BUCKET_NAME,
+            Key: fileName,
+            Body: file.buffer,
+            ContentType: file.mimetype
+        };
+
+        await s3.send(new PutObjectCommand(params));
+        const imageUrl = `https://${process.env.AWS_BUCKET_NAME}.s3.${process.env.AWS_REGION}.amazonaws.com/${fileName}`;
+
+        const bookId = req.params.id;
+
+        if(!validator.isUUID(bookId)){
+            return res.status(400).json({
+                message: "Invalid Book ID"
+            })
+        }
+
+        const result = await query(
+            `SELECT * from ebooks where book_id=$1 and user_id=$2`,
+            [bookId,req.user.id]
+        )
+        if(result.rows.length===0){
+            return res.status(404).json({
+                message: "Book not found"
+            })
+        }
+
+        const book = await query(
+            `UPDATE ebooks set cover_image_url=$1,cover_image_key=$2 where book_id=$3 and user_id=$4 RETURNING *`,
+            [imageUrl,fileName,bookId,req.user.id]
+        );
+
+        return res.status(201).json({
+            message: "Successfully added cover image!",
+            book: book.rows[0]
+        })
 
     } catch(err){
         return res.status(500).json({
@@ -39,8 +88,9 @@ async function addCover(req,res){
 
 async function getBooks(req,res){
     try{
+        console.log(req.user.id);
         const result = await query(
-            `SELECT * FROM ebooks`);
+            `SELECT * FROM ebooks where user_id=$1`,[req.user.id]);
         return res.status(200).json({
             message: "Fetched all the EBooks! ",
             books: result.rows
@@ -61,8 +111,8 @@ async function updateEbook(req,res){
             })
         }
         const book = await query(
-            `select * from ebooks where book_id=$1`,
-            [bookId]
+            `select * from ebooks where book_id=$1 and user_id=$2`,
+            [bookId,req.user.id]
         );
 
         if(book.rows.length===0){
@@ -71,8 +121,8 @@ async function updateEbook(req,res){
             })
         }
         const result = await query(
-            `UPDATE ebooks set book_title=$1,cover_image_url=$2,description=$3 where book_id=$4 RETURNING *`,
-            [title,url,description,bookId]
+            `UPDATE ebooks set book_title=$1,cover_image_url=$2,description=$3 where book_id=$4 and user_id=$5 RETURNING *`,
+            [title,url,description,bookId,req.user.id]
         );
         return res.status(200).json({
             message:"Updated title of the ebook",
@@ -95,8 +145,8 @@ async function deleteEbook(req,res){
             })
         }
         const book = await query(
-            `select * from ebooks where book_id=$1`,
-            [bookId]
+            `select * from ebooks where book_id=$1 and user_id=$2`,
+            [bookId,req.user.id]
         );
 
         if(book.rows.length===0){
@@ -105,8 +155,8 @@ async function deleteEbook(req,res){
             })
         }
         const result = await query(
-            `DELETE FROM ebooks where book_id=$1`
-            ,[bookId]);
+            `DELETE FROM ebooks where book_id=$1 and user_id=$2`
+            ,[bookId,req.user.id]);
         
         return res.status(200).json({
             message: "Deleted an E-Book",
@@ -119,4 +169,4 @@ async function deleteEbook(req,res){
         })
     }
 }
-export default {createBook,getBooks,updateEbook,deleteEbook};
+export default {createBook,getBooks,updateEbook,deleteEbook,addCover};
